backend/requirements.txt
# AWS SDK
boto3>=1.34.0

# API Framework
fastapi>=0.109.0
mangum>=0.17.0

# AI/ML
anthropic>=0.8.0
openai>=1.7.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.10.0
requests>=2.31.0

# JWT Authentication
python-jose[cryptography]>=3.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
moto>=5.0.0
